Fix wrapping with 3 lines or more
The current line length is set to 0 from the second line while it should
be set to the current word length plus space instead.

Moreover, the theoritical number of lines is computed too naively: it
does not take wrapping constraints into consideration. Instead, compute
the number of line breaking when trying to wrap in a classic way.